file dialog created and folder dialog renamed appropriately
