+++ src/modules/dialogs/folder_dialog.h
@@ -1,0 +1,4 @@
+#ifndef FILE_DIALOG_H
+#define FILE_DIALOG_H
+
+#endif // FILE_DIALOG_H
